financesvc/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
drf-yasg==1.21.7
django-filter==24.1
gunicorn 